Updated mylyn/org.eclipse.mylyn.all Project: mylyn/org.eclipse.mylyn  d9c37d0bda12f5e5c541f5447c169d83d825a635
423916 - Fixed ID of Tuleap feature.

"org.tuleap" was not a correct ID to be used, we hadn't understood that
it had to match the feature ID in the update site.
For the moment, the referenced update site has been temporarily modified
to use the ID "org.tuleap", but as soon as this change is merged and
deployed we'll get back to "org.tuleap.mylyn.task-feature" as this is
what the Tuleap connector automated build produces.

Change-Id: Iba28578b8c85e5014ab0723b442daf7b6c62d252
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>