Updated mylyn/org.eclipse.mylyn.all Project: mylyn/org.eclipse.mylyn.commons  ab4063a60bb1345e3f54a6616bb956a0a4c6c328
495085: connector discovery should tolerate servers that return 403 for
non-existent resources

Discovery looks for any of "content.jar", "content.xml",
"compositeContent.jar", "compositeContent.xml", "site.xml" to determine
if an update site is valid. It assumes it will get a 404 for resources
that don't exist. But it is common for servers (e.g. AWS) to use 403 in
that case. It should succeed if any of the resources exist, regardless
of the error that occurs accessing the others.

Change-Id: I2c4e0a2000e59672071ce8f30822ad54110cb763
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=495085