Updated mylyn/org.eclipse.mylyn.all Project: mylyn/org.eclipse.mylyn.tasks  49c3cb0bfe61796f484b52c629a98172bb7eefee
replace testValidateOnFinishInvalidUserId with unit test

This test is running forever in http://ci.mylyn.org/job/mylyn-3.21.x/
and I can't figure out why. This is an integration test whose main
purpose is to verify that the settings page remains open when validate
on finish fails. This really ought to be verified by a unit test of the
page instead.

Change-Id: I55813566e3707f082904001750c065285ed5c3e0