Updated mylyn/org.eclipse.mylyn.all Project: mylyn/org.eclipse.mylyn.tasks  d2a46aff5af616b1d97e24f1702e856234ca76e7
378863: support replying to selected part of a comment only

If the comment viewer is a styled text widget the reply to action will
attempt to reply to only the selected comment text. This means that some
connectors which use other controls for their comments will not support
this functionality, and the editor will default to replying ot the
entire comment text.

Change-Id: Ie100b62e4405da0720c3655e01fd18e6f1a181f0
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=378863